backend/requirements.txt
Django>=2.0,<3.0
psycopg2==2.8.6
djangorestframework
djangorestframework-jwt
pillow
django-cors-headers